+++ project/src/components/genre-list/genre-list.jsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import PropTypes from 'prop-types'
+import PropTypes from 'prop-types';
 import {connect} from 'react-redux';
 import Genre from '../genre/genre.jsx';
 import {ActionCreator} from '../../store/action.js';
