+++ project/src/components/my-list/my-list.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import {connect} from 'react-redux';
-import HeaderWithAuthBlock from '../header-with-auth-block/header-with-auth-block.jsx';
+import AuthHeader from '../auth-header/auth-header.jsx';
 import Footer from '../footer/footer.jsx';
 import FilmList from '../film-list/film-list.jsx';
 import {selectFavoriteList} from '../../selectors/selectors.js';
@@ -10,9 +10,9 @@ import {filmsPropTypes} from '../../prop-types/films.js';
 function MyList({films}) {
   return (
     <div className="user-page">
-      <HeaderWithAuthBlock specialClass={HeaderClass.USER_PAGE}>
+      <AuthHeader specialClass={HeaderClass.USER_PAGE}>
         <h1 className="page-title user-page__title">My list</h1>
-      </HeaderWithAuthBlock>
+      </AuthHeader>
 
       <section className="catalog">
         <h2 className="catalog__title visually-hidden">{CatalogTitle.MY_LIST}</h2>
